Feat: Update Logout alert() to Alert modal
